get rid of multiple #if with polyfill and remove serialization
