+++ src/NMoneys/NMoneys.csproj
@@ -52,17 +52,18 @@
     <Compile Include="Allocations\RatioCollection.cs" />
     <Compile Include="Allocations\RemainderAllocatorBase.cs" />
     <Compile Include="Allocations\LastToFirstAllocator.cs" />
-    <Compile Include="Currency.Serialization.Xml.cs" />
-    <Compile Include="Currency.Serialization.cs" />
+    <Compile Include="Currency.Serialization.Xml.net.cs" />
+    <Compile Include="Currency.Serialization.net.cs" />
     <Compile Include="Currency.Code.cs" />
     <Compile Include="Currency.Creation.cs" />
     <Compile Include="Currency.Formatting.cs" />
     <Compile Include="Currency.Comparisons.cs" />
     <Compile Include="Currency.Equality.cs" />
     <Compile Include="CurrencyCache.cs" />
+    <Compile Include="Exceptions.net.cs" />
     <Compile Include="Money.Creation.cs" />
-    <Compile Include="Money.Serialization.Xml.cs" />
-    <Compile Include="Money.Serialization.cs" />
+    <Compile Include="Money.Serialization.Xml.net.cs" />
+    <Compile Include="Money.Serialization.net.cs" />
     <Compile Include="Money.Arithmetic.cs" />
     <Compile Include="Money.Allocations.cs" />
     <Compile Include="Money.Arithmetic.Operators.cs" />
