+++ src/NMoneys/Support/Serialization.polyfill.netstandard.cs
@@ -1,0 +1,9 @@
+﻿namespace System.Xml.Serialization
+{
+    public class XmlIgnoreAttribute : Attribute { }
+}
+
+namespace System.Runtime.Serialization
+{
+	public class EnumMemberAttribute : Attribute { }
+}
